+++ packages/enhanced/jest.config.ts
@@ -36,7 +36,6 @@ export default {
   testMatch: [
     '<rootDir>/test/*.basictest.js',
     '<rootDir>/test/unit/**/*.test.ts',
-    '<rootDir>/test/compiler-unit/**/*.test.ts',
   ],
 
   testEnvironment: path.resolve(__dirname, './test/patch-node-env.js'),