+++ packages/nextjs-mf/src/plugins/NextFederationPlugin/flight-client-entry-plugin.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { normalizeWebpackPath } from '@module-federation/sdk/normalize-webpack-path';
 
 import type {