+++ packages/rsbuild-plugin/src/utils/ssr.ts
@@ -1,9 +1,13 @@
 import path from 'path';
+import { createRequire } from 'node:module';
 import { encodeName } from '@module-federation/sdk';
+import { CALL_NAME_MAP } from '../constant';
 
 import type { EnvironmentConfig, RsbuildConfig, Rspack } from '@rsbuild/core';
 import type { moduleFederationPlugin } from '@module-federation/sdk';
-import { CALL_NAME_MAP } from '../constant';
+
+const require = createRequire(import.meta.url);
+const resolve = require.resolve;
 
 export const SSR_DIR = 'ssr';
 export const SSR_ENV_NAME = 'mf-ssr';
@@ -113,15 +117,16 @@ export function createSSRMFConfig(
     },
     dts: false,
     dev: false,
+    runtimePlugins: [...(mfConfig.runtimePlugins || [])],
   };
 
-  ssrMFConfig.runtimePlugins ||= [];
   ssrMFConfig.runtimePlugins.push(
-    require.resolve('@module-federation/node/runtimePlugin'),
+    resolve('@module-federation/node/runtimePlugin'),
   );
   if (isDev()) {
     ssrMFConfig.runtimePlugins.push(
-      require.resolve(
+      // @ts-ignore
+      resolve(
         '@module-federation/node/record-dynamic-remote-entry-hash-plugin',
       ),
     );