+++ packages/bridge/bridge-shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @module-federation/bridge-shared
 
+## 0.16.0
+
 ## 0.15.0
 
 ## 0.14.3