+++ packages/bridge/vue3-bridge/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/bridge-vue3
 
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies [1485fcf]
+  - @module-federation/sdk@0.16.0
+  - @module-federation/runtime@0.16.0
+  - @module-federation/bridge-shared@0.16.0
+
 ## 0.15.0
 
 ### Patch Changes