+++ packages/modernjs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @module-federation/modern-js
 
+## 0.16.0
+
+### Patch Changes
+
+- 98136ca: fix(modern-js-plugin): use contenthash instead of chunkhash
+- de350f3: fix(modern-js-plugin): adjust fetch type
+- Updated dependencies [1485fcf]
+- Updated dependencies [98136ca]
+- Updated dependencies [98136ca]
+  - @module-federation/sdk@0.16.0
+  - @module-federation/node@2.7.8
+  - @module-federation/rsbuild-plugin@0.16.0
+  - @module-federation/cli@0.16.0
+  - @module-federation/enhanced@0.16.0
+  - @module-federation/runtime@0.16.0
+
 ## 0.15.0
 
 ### Minor Changes