+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @module-federation/utilities
 
+## 3.1.60
+
+### Patch Changes
+
+- Updated dependencies [1485fcf]
+  - @module-federation/sdk@0.16.0
+
 ## 3.1.59
 
 ### Patch Changes